RavenDB-26110 Refactor growable buffer initialization.
